% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepforest_utils.R
\name{getFolds}
\alias{getFolds}
\title{Get n-folds}
\usage{
getFolds(y, nfold = 5, recycle = FALSE, list = TRUE)
}
\description{
Shamelessly taken from \code{\link{caret::createFolds}}
https://topepo.github.io/caret/
}
