% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepforest_utils.R
\name{metaparam}
\alias{metaparam}
\title{Meta Model Parameters}
\usage{
metaparam(metatype, param = rep(list(list()), length(metatype)),
  metarandom = FALSE)
}
\description{
Used for setting default meta model parameters
}
