% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_cv.R
\name{ranger_cv}
\alias{ranger_cv}
\title{random forest cv}
\usage{
ranger_cv(i, formula, data, num.trees, folds, mtry, splitrule, probability,
  write.forest, num.threads, save.memory)
}
\arguments{
\item{i}{i-esimo fold}

\item{formula}{formula del modello}

\item{folds}{prodotto da kfold function}

\item{mtry}{}

\item{y}{vettore. variabile risposta}

\item{x}{matrice variabili esplicative}
}
\description{
fonisce le previsioni della cv, da utilizzare dentro sapply(1:k,...)
}
